Latitude and longitude are reversed
The JSON output from Timezone Boundary Builder awkwardly encodes its
coordinates in float arrays of longitude and latitude, rather than the
other way around, which would be closer to the spoken English
convention.  We need to decode the file with that in mind, and then fix
up all the examples which had been reversed to get the right answer.

Also constrain the latitude from -90 to 90 degrees.