Fix mmap/munmap to use portable calls
Use the package "github.com/edsrzf/mmap-go" to perform
os independent mmap and munmap calls.  As a result of this
change another platform dependent was discovered because of
using `pageSize` as an offset.  On Windows, the MapView call
requires an offset that is not a multiple of `pageSize`, but
is a multiple of the `AllocationGranularity` which is itself
a multiple of `pageSize`.  The easiest way to get around
this is to drop the idea of using offsets in the mmap and
just start at 0.  This means that the total allocated mmap
is larger by the size of the offset but that seems to be
a small price to pay to keep the code simple and portable.